feat: generate kubernetes_secrets manifests (#6)
To simplify the getting started guide, I've added a new output called `kubernetes_secrets` that can be piped to `kubectl apply` to create the secrets. This just allows us to make the guide a little simpler, while still giving people the ability to take a look under the hood if they want.

In order to generate the secret yaml definitions, we need some extra information that wasn't previously provided. All of these variables are optional, and can be ignored if not following the guide.

I also adjusted the template variables to use `=` instead of `:` for assignment so that the code formats correctly.