feat: simplify module to use EKS auto (#11)
The reason I hadn't originally used EKS auto was that I wanted to use Security Groups per Pod to keep the security setup as close to how we'd typically deploy Spacelift, but it's really not worth it for the restrictions it adds and the additional complexity.

Given that this module is intended as an example to go along with our getting started docs, I want to simplify it for now. Customers can still choose to configure things differently if they want, but we'll keep our example a bit simpler for now.